Deleted a few more no-longer used scripts.
Heavily re-wrote mdcheck.  Uses cut rather than awk for efficiency;
now calls the checking pattern from inside the if-loop checking that
the file exists, eliminating the clunky 123 >/dev/null as a way of
breaking out of the loop.

Beginning to modify shell files so that they use sh rather than bash
where possible, and all adhere to a consistent style.